cilium: extend IPSec to allow programming FWD policy and template IPs
So far we've programmed FWD policy rules to mirror the IN policy rules,
but this requires the outer IPSec tunnel dstIP to match the inner
dstIP. Also we've been setting the template IPs the same as the matching
src/dst IPs. We are about to fix the ENI case where this is not always
true.

In order to support this allow pushing uniue FWD address and push
template IPs.

Note: It probably makes sense to do a heavier refactor here and rip
out UpsertIPsecEndpoint(), at this point it just adds confusion in
my opinion. But, lets avoid that at the moment for backports.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>