cilium: xfrm netlink, we need to program the optional field
The optional field will instruct the datapath to drop the packet
if the tmpl lookup fails. In our case its we have Cilium to ensure
encrypted/unencrypted packets are correctly managed we don't need
the stack to get in the way.

This patch updates netlink code to accept the option.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>